# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IAM Platform API for EduGo - authentication, roles, permissions, resources, menu and screen configuration",
        "title": "EduGo API IAM Platform",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8070",
    "basePath": "/api/v1",
    "paths": {
        "/auth/contexts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return all school contexts the authenticated user has access to",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get available school contexts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.AvailableContextsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticate a user and return access/refresh tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login with email and password",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalidate the current access token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Use a refresh token to get new access/refresh tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh access token",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/switch-context": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Switch the user's active school context and get new tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Switch school context",
                "parameters": [
                    {
                        "description": "Target school ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.SwitchContextRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/verify": {
            "post": {
                "description": "Validate a JWT token and return its claims",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify JWT token",
                "parameters": [
                    {
                        "description": "Token to verify",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.VerifyTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.VerifyTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.AvailableContextsResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.UserContextDTO"
                    }
                },
                "current": {
                    "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.UserContextDTO"
                }
            }
        },
        "github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "active_context": {
                    "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.UserContextDTO"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "schools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.SchoolInfo"
                    }
                },
                "token_type": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.UserInfo"
                }
            }
        },
        "github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.RefreshTokenRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.SchoolInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.SwitchContextRequest": {
            "type": "object",
            "required": [
                "school_id"
            ],
            "properties": {
                "school_id": {
                    "type": "string"
                }
            }
        },
        "github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.UserContextDTO": {
            "type": "object",
            "properties": {
                "academic_unit_id": {
                    "type": "string"
                },
                "academic_unit_name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "role_id": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "school_name": {
                    "type": "string"
                }
            }
        },
        "github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.UserInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                }
            }
        },
        "github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.VerifyTokenRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_EduGoGroup_edugo-api-iam-platform_internal_auth_dto.VerifyTokenResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token. Example: \"Bearer eyJhbGci...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}